+++ test_runner/src/main/kotlin/ftl/args/yml/FlankYml.kt
@@ -2,41 +2,62 @@ package ftl.args.yml
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
+import ftl.args.ArgsHelper
+import ftl.config.FtlConstants
 import ftl.config.FtlConstants.GCS_PREFIX
+import ftl.util.Utils
 import ftl.util.Utils.fatalError
 
 /** Flank specific parameters for both iOS and Android */
 @JsonIgnoreProperties(ignoreUnknown = true)
 class FlankYmlParams(
+    @field:JsonProperty("max-test-shards")
     val maxTestShards: Int = 1,
+
+    @field:JsonProperty("shard-time")
     val shardTime: Int = -1,
+
+    @field:JsonProperty("repeat-tests")
     val repeatTests: Int = 1,
+
+    @field:JsonProperty("smart-flank-gcs-path")
     val smartFlankGcsPath: String = "",
+
+    @field:JsonProperty("disable-sharding")
     val disableSharding: Boolean = false,
 
     @field:JsonProperty("test-targets-always-run")
     val testTargetsAlwaysRun: List<String> = emptyList(),
 
     @field:JsonProperty("files-to-download")
-    val filesToDownload: List<String> = emptyList()
+    val filesToDownload: List<String> = emptyList(),
+
+    val project: String = ArgsHelper.getDefaultProjectId() ?: ""
 ) {
     companion object : IYmlKeys {
         override val keys = listOf(
-            "maxTestShards", "shardTime", "repeatTests", "smartFlankGcsPath", "disableSharding", "test-targets-always-run", "files-to-download"
+            "max-test-shards", "shard-time", "repeat-tests", "smart-flank-gcs-path", "disable-sharding",
+            "test-targets-always-run", "files-to-download", "project"
         )
     }
 
     init {
-        if (maxTestShards <= 0 && maxTestShards != -1) fatalError("maxTestShards must be >= 1 or -1")
-        if (shardTime <= 0 && shardTime != -1) fatalError("shardTime must be >= 1 or -1")
-        if (repeatTests < 1) fatalError("repeatTests must be >= 1")
+        Utils.assertNotEmpty(
+            project, "The project is not set. Define GOOGLE_CLOUD_PROJECT, set project in flank.yml\n" +
+                    "or save service account credential to ${FtlConstants.defaultCredentialPath}\n" +
+                    " See https://github.com/GoogleCloudPlatform/google-cloud-java#specifying-a-project-id"
+        )
+
+        if (maxTestShards <= 0 && maxTestShards != -1) fatalError("max-test-shards must be >= 1 or -1")
+        if (shardTime <= 0 && shardTime != -1) fatalError("shard-time must be >= 1 or -1")
+        if (repeatTests < 1) fatalError("repeat-tests must be >= 1")
 
         if (smartFlankGcsPath.isNotEmpty()) {
             if (!smartFlankGcsPath.startsWith(GCS_PREFIX)) {
-                fatalError("smartFlankGcsPath must start with gs://")
+                fatalError("smart-flank-gcs-path must start with gs://")
             }
             if (smartFlankGcsPath.count { it == '/' } <= 2 || !smartFlankGcsPath.endsWith(".xml")) {
-                fatalError("smartFlankGcsPath must be in the format gs://bucket/foo.xml")
+                fatalError("smart-flank-gcs-path must be in the format gs://bucket/foo.xml")
             }
         }
     }