+++ test_runner/src/main/kotlin/ftl/args/yml/YamlDeprecated.kt
@@ -1,6 +1,7 @@
 package ftl.args.yml
 
 import com.fasterxml.jackson.databind.JsonNode
+import com.fasterxml.jackson.databind.node.JsonNodeFactory
 import com.fasterxml.jackson.databind.node.ObjectNode
 import ftl.args.ArgsHelper.yamlMapper
 import ftl.util.Utils
@@ -29,8 +30,33 @@ object YamlDeprecated {
     private val transforms = listOf(
         // flank: testShards -> flank: maxTestShards
         ModifiedKey(
-            Key(YamlDeprecated.Parent.flank, "testShards"),
-            Key(YamlDeprecated.Parent.flank, "maxTestShards"),
+            Key(Parent.flank, "testShards"),
+            Key(Parent.flank, "max-test-shards"),
+            Level.Warning
+        ),
+        ModifiedKey(
+            Key(Parent.flank, "shardTime"),
+            Key(Parent.flank, "shard-time"),
+            Level.Warning
+        ),
+        ModifiedKey(
+            Key(Parent.flank, "repeatTests"),
+            Key(Parent.flank, "repeat-tests"),
+            Level.Warning
+        ),
+        ModifiedKey(
+            Key(Parent.flank, "smartFlankGcsPath"),
+            Key(Parent.flank, "smart-flank-gcs-path"),
+            Level.Warning
+        ),
+        ModifiedKey(
+            Key(Parent.flank, "disableSharding"),
+            Key(Parent.flank, "disable-sharding"),
+            Level.Warning
+        ),
+        ModifiedKey(
+            Key(Parent.gcloud, "project"),
+            Key(Parent.flank, "project"),
             Level.Warning
         )
     )
@@ -45,7 +71,10 @@ object YamlDeprecated {
     }
 
     private fun JsonNode.replace(parent: Parent, child: String, value: JsonNode) {
-        (this[parent.toString()] as ObjectNode).replace(child, value)
+        val parentKey = parent.toString()
+        // if the parent node ('flank:') doesn't exist then add it ('flank: {}') to the YAML
+        if (this[parentKey] == null) (this as ObjectNode).set(parentKey, JsonNodeFactory.instance.objectNode())
+        (this[parentKey] as ObjectNode).replace(child, value)
     }
 
     private fun mutate(parsed: JsonNode, changes: List<Transform>) {